Add GetPhotosRequestBuilder and Fix Layout
- use a builder to construct the request
- pass the builder to the api manager to build the request, run it,
  return the response
- update the layouts to remove the padding on the top / bottom of the
  recycler view

+++ android/app/src/main/java/xp/app/go500px/PhotosAdapter.java
@@ -26,6 +26,6 @@ public class PhotosAdapter extends RecyclerView.Adapter<PhotoViewHolder> {
 
   @Override
   public int getItemCount() {
-    return (int) mPhotos.Count();
+    return mPhotos.Count();
   }
 }
